requirements.txt
APScheduler==3.0.5
Django==1.8.5
Pillow==3.0.0
decorator==4.0.4
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.23.0
django-bootstrap3==6.2.2
django-configurations==0.8
django-toolbelt==0.0.1
djangorestframework==3.3.2
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
python3-openid==3.0.7
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.5.0
shell==1.0.1
simplegeneric==0.8.1
six==1.10.0
static3==0.6.1
traitlets==4.0.0
tzlocal==1.2
urllib3==1.14
